requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.8.2
alembic==1.13.1
